pyproject.toml
[project]
name = "energnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "ipykernel",
    "numpy",
    "optimade",
    "pymatgen",
    "pymatviz",
    "pysnooper",
    "tqdm",
    "ultraimport",
    "wandb",
    "torch",
    "torch-scatter",
    "fairchem-core",
    "torch_geometric",
    "monty",
    "matbench-discovery@git+https://github.com/janosh/matbench-discovery.git"
]
[tool.uv.extra-build-dependencies]
torch-scatter=["torch"]
[[tool.uv.index]]
name="pytorch-cu126"
url="https://download.pytorch.org/whl/cu126"
explicit = true
[tool.uv.sources]
torch=[
    {index="pytorch-cu126", marker="sys_platform=='linux' or sys_platform=='win32'"}
]
torchvision = [
    {index="pytorch-cu126", marker="sys_platform=='linux' or sys_platform=='win32'"}
]
